+++ internal/completion/helpers.go
@@ -1,7 +1,6 @@
 package completion
 
 import (
-	"fmt"
 	"strings"
 )
 
@@ -9,22 +8,12 @@ import (
 func MatchStringPrefix(vals []string, key string, caseSensitive bool) []string {
 	var r []string
 	key = strings.Trim(key, "'\"")
-
 	for _, v := range vals {
 		if (caseSensitive && strings.HasPrefix(v, key)) ||
 			(!caseSensitive && strings.HasPrefix(strings.ToLower(v), strings.ToLower(key))) ||
 			key == "" {
-			r = append(r, Escape(v))
+			r = append(r, v)
 		}
 	}
 	return r
-}
-
-// Escape escapes a string according to completion rules (?)
-// in effect, this means that the string will be quoted with double quotes if it contains a space or parentheses.
-func Escape(s string) string {
-	if strings.ContainsAny(s, ` ()`) {
-		return fmt.Sprintf(`"%s"`, s)
-	}
-	return s
 }