+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added support for all shell completions provided by `cobra`.
+
+### Changed
+- Remove custom bash completion logic and replace it with `completion` command provided by `cobra`.
 
 ## [1.5.0] - 2022-07-05
 ### Added